Update README.md
This text outlines an intensive four-month learning plan for transitioning into a Strategic DevOps Engineer role. The plan emphasizes daily study and practical application through six diverse projects. It also highlights building a strong presence on GitHub and Docker Hub, alongside developing an advanced research paper on Autoscaling. The ultimate goal is to acquire the necessary skills and experience for a promising career in DevOps.